% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{calcGOenrichment}
\alias{calcGOenrichment}
\title{Perform GO enrichment analysis}
\usage{
calcGOenrichment(
  goAnnotation,
  deGenes = character(),
  universe = character(),
  domain = "BP",
  minGenes = 5,
  k = 1,
  nCores = 0
)
}
\arguments{
\item{goAnnotation}{evoGO object or list of evoGO objects}

\item{deGenes}{character vector of differentially expressed genes}

\item{universe}{character vector of all possible genes for this dataset}

\item{domain}{character, name of GO domain: 'BP' for biological process, "CC" for
cellular component, "MF" for molecular function.}

\item{minGenes}{integer indicating the minimal number of genes attributed to
a GO term for this term to be considered during the analysis. This value should
not be lower then 3. Recommended value: 5.}

\item{k}{numeric, additional coefficient to be applied to weights when
down-weighting genes}

\item{nCores}{integer, maximum number of cores used during calculations.
If 0, maximum available number of cores is used but not more than 8.}
}
\value{
data.frame containing the following columns:
\itemize{
  \item{id - GO term ID}
  \item{name - GO term name}
  \item{def - GO term definition}
  \item{fisher.pvalue - p value obtained using classic overrepresentation analysis 
  relying on Fisher's exact test}
  \item{evogo.pvalue - p value provided by evoGO algorithm}
  \item{annotated - number of genes attributed to a corresponding GO term}
  \item{significant - number of differentially expressed genes for a corresponding GO term}
}
}
\description{
Perform GO enrichment analysis
}
