% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{evoGO-package}
\alias{evoGO-package}
\alias{_PACKAGE}
\title{evoGO: GO enrichment analysis using GO term weighting algorithm}
\description{
Please refer to the README and function documentation to learn more
about the usage.
}
\author{
\strong{Maintainer}: Ievgen Strielkov \email{github_bioinfo@evotec.com}

Authors:
\itemize{
  \item Vladimir Galatenko
}

Other contributors:
\itemize{
  \item Alexey Uvarovskii [contributor]
}

}
\keyword{internal}
