% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go-resources.R
\name{exampleGOAnnotation}
\alias{exampleGOAnnotation}
\title{Download example annotation for Homo sapiens}
\usage{
exampleGOAnnotation(path = NULL)
}
\arguments{
\item{path}{directory where example evoGO annotation will be stored. If NULL, the default
package directory is used ('extdata').}
}
\value{
list of evoGO objects containing the example annotation (one per GO domain).
}
\description{
This function obtains an example annotation and saves it to the package 'extdata'
directory (a default directory for stored annotations). The purpose of the function is to
provide the means for quick testing of the package functions. For performing a proper GO
enrichment analysis please acquire an up-to-date annotation using \code{\link{getGOAnnotation}}.
}
