% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{getCoefs}
\alias{getCoefs}
\title{Helper function that calculates coefficients for downweighting of genes of
the current node (term) in the ancestor nodes}
\usage{
getCoefs(go, goWeights, goP, k = 1, g)
}
\arguments{
\item{go}{character, GO term ID}

\item{goWeights}{numeric, named vector containing gene weights for the
current term}

\item{goP}{numeric, p value of the current term}

\item{k}{numeric, additional coefficient to be applied to weights when
downweighting genes}

\item{g}{environment containing all nodes}
}
\value{
list of coefficients for genes of the ancestor terms
}
\description{
Helper function that calculates coefficients for downweighting of genes of
the current node (term) in the ancestor nodes
}
