% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go-resources.R
\name{getCustomGOAnnotation}
\alias{getCustomGOAnnotation}
\title{Construct and save a GO annotation using a cutom gene-GO mapping}
\usage{
getCustomGOAnnotation(
  customAnnotation = NULL,
  customName = NULL,
  goRelease = NULL,
  nCores = 1,
  save = TRUE,
  path = NULL,
  deletePrevious = FALSE
)
}
\arguments{
\item{customAnnotation}{a named list with character vectors of gene groups per GO term,
e.g. list(`GO:01` = c("ENS01", "ENS02"), `GO:02` = c("ENS03")), where
the names are the GO term identifiers.
The list items may contain the genes belonging to all children terms or
include only the genes with exclusion of the children terms.
See output of getGeneSets() as format reference.}

\item{customName}{a string that is used to store custom annotation file.}

\item{goRelease}{GO release date in yyyy-mm-dd format (e.g., "2022-01-31").
If not specified, latest version is retrieved.}

\item{nCores}{maximum number of cores to be used.}

\item{save}{logical value indicating whether the result should be saved (default TRUE).}

\item{path}{path to a directory where the result will be saved. If NULL, the file
will be saved to the package 'extdata' directory.}

\item{deletePrevious}{logical value indicating whether previous annotations for
that species should be removed when newer is saved to save disk space (default FALSE).}
}
\value{
list of evoGO objects containing the annotation (one per GO domain).
}
\description{
Construct and save a GO annotation using a cutom gene-GO mapping
}
