% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go-resources.R
\name{getGOAnnotation}
\alias{getGOAnnotation}
\title{Retrieve and save the latest annotation for species of interest}
\usage{
getGOAnnotation(
  species,
  database = "ensembl",
  nCores = 1,
  save = TRUE,
  path = NULL,
  deletePrevious = TRUE
)
}
\arguments{
\item{species}{species identifier (example: 'hsapiens' for human)}

\item{database}{database to be used for acquiring of genesets. Only 'ensembl'
is currently supported.}

\item{nCores}{maximum number of cores to be used}

\item{save}{logical value indicating whether the result should be saved}

\item{path}{path to a directory where the result will be saved. If NULL, the file
will be saved to the package 'extdata' directory}

\item{deletePrevious}{logical value indicating whether previous annotation for
that species should be removed when newer is saved to save disk space. Only
used in case results are saved to 'extdata' directory}
}
\value{
list of evoGO objects containing the latest annotation (one per GO domain).
}
\description{
Retrieve and save the latest annotation for species of interest
}
