% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{getGOscores}
\alias{getGOscores}
\title{Function that downweights genes of the given term in the less significant
ancestor terms and calculates resulting p values for the whole GO term graph}
\usage{
getGOscores(data, k = 1, nCores = 1)
}
\arguments{
\item{data}{object containing GO graph}

\item{k}{numeric, additional coefficient to be applied to weights when
downweighting genes}

\item{nCores}{integer, maximum number of processes spawned during
calculations}
}
\value{
updated graph object
}
\description{
Function that downweights genes of the given term in the less significant
ancestor terms and calculates resulting p values for the whole GO term graph
}
