% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go-resources.R
\name{getOBO}
\alias{getOBO}
\title{Retrieve the ontologyIndex object from the obolibrary.}
\usage{
getOBO()
}
\value{
ontologyIndex object
}
\description{
Retrieve the ontologyIndex object from the obolibrary.
}
