% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gtex_heart_brain}
\alias{gtex_heart_brain}
\title{Differential gene expression: Heart (Left Ventricle) vs. Brain (Cortex)}
\format{
A data frame with 36697 rows and 5 variables:
\describe{
  \item{GeneID}{Ensembl gene ID}
  \item{CPM}{counts per million, average of the normalized expression values}
  \item{log2FoldChange}{log2(FoldChange), effect size estimate}
  \item{Pvalue}{P value}
  \item{Pvalue.adjusted}{P value adjusted using Benjamini-Hochberg procedure}
}
}
\source{
\url{https://gtexportal.org/}
}
\usage{
gtex_heart_brain
}
\description{
Human tissue gene expression data was obtained from GTEx V8 database. Only 
samples with tissue degradation level 0 or 1 and death circumstances 1-3 on 
4-point Hardy Scale were used.
In total, 29 heart samples and 19 brain samples were used for the analysis.
}
\keyword{datasets}
