% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go-resources.R
\name{listGOAnnotations}
\alias{listGOAnnotations}
\title{List previously stored GO annotation files}
\usage{
listGOAnnotations(
  species = NULL,
  goRelease = NULL,
  ensemblRelease = NULL,
  customName = NULL,
  database = "ensembl",
  path = NULL,
  returnLatest = FALSE
)
}
\arguments{
\item{species}{species identifier (example: 'hsapiens' for human).}

\item{goRelease}{GO release date in yyyy-mm-dd format (e.g., "2022-01-31").}

\item{ensemblRelease}{numeric gene annotation version (e.g., 100 for Ensembl).}

\item{customName}{a string that was used to save custom annotation file.}

\item{database}{database used for acquiring of genesets. Only 'ensembl' is
currently supported.}

\item{path}{path to the directory where the annotation files were saved.}

\item{returnLatest}{logical value indicating whether to return information for
the file with the latest GO and Ensembl releases (default FALSE). If the list of
files contains both standard and custom annotation files, the standard annotation
file with the latest GO and Ensembl releases will be selected as custom annotation
files do not use the Ensembl database.}
}
\value{
a data.frame containing information about previously stored GO annotations.
}
\description{
List of available files can be filtered using filter arguments.
}
