% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go-resources.R
\name{loadGOAnnotation}
\alias{loadGOAnnotation}
\title{Load previously retrieved GO annotation as evoGO object}
\usage{
loadGOAnnotation(
  species,
  database = "ensembl",
  goRelease = NULL,
  annoVersion = NULL,
  path = NULL
)
}
\arguments{
\item{species}{species identifier (example: 'hsapiens' for human)}

\item{database}{database to be used for acquiring of genesets. Only 'ensembl' is
currently supported}

\item{goRelease}{GO release date in yyyy-mm-dd or yyyymmdd format (e.g., "2022-01-31").
If not specified, latest saved version is loaded}

\item{annoVersion}{gene annotation version (e.g., "GRCh38.p13" for Ensembl). If not specified,
latest saved version is loaded}

\item{path}{directory where evoGO annotations are stored. If NULL, the default package
directory is used ('extdata')}
}
\value{
list of evoGO objects containing the latest annotation (one per GO domain).
}
\description{
Load previously retrieved GO annotation as evoGO object
}
