% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go-resources.R
\name{ontologyIndex2graph}
\alias{ontologyIndex2graph}
\title{Create an input data.frame for evoGO constructor function}
\usage{
ontologyIndex2graph(go, rootName = "cellular_component")
}
\arguments{
\item{go}{an ontologyIndex object returned by getOBO function (also see ontologyIndex package).}

\item{rootName}{the name of the root elements from which to build
the hierarchy.}
}
\value{
a data.frame to be used as graphTable argument in evoGO function.
}
\description{
The data.frame includes "child" - "parent" relationships between the GO ids.
}
