% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{setDEGs}
\alias{setDEGs}
\title{Helper function that assigns differentially expressed genes to graph nodes
(updates 'degs' node attribute)}
\usage{
setDEGs(diffExprIDs, data, nCores = 1)
}
\arguments{
\item{diffExprIDs}{character vector of gene IDs}

\item{data}{object containing GO graph}

\item{nCores}{maximum number of processes}
}
\value{
updated data object containing GO graph
}
\description{
Helper function that assigns differentially expressed genes to graph nodes
(updates 'degs' node attribute)
}
