% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{setInitPVals}
\alias{setInitPVals}
\title{Helper function that calculates (using one-sided Fisher exact test) and
assigns p values to nodes (updates 'p' node attribute)}
\usage{
setInitPVals(data, nCores = 1)
}
\arguments{
\item{data}{object containing GO graph}

\item{nCores}{maximum number of processes}
}
\value{
updated igraph object containing GO graph
}
\description{
Helper function that calculates (using one-sided Fisher exact test) and
assigns p values to nodes (updates 'p' node attribute)
}
