% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{setUniverse}
\alias{setUniverse}
\title{Helper function that removes genes from graph nodes based on the given set of
possible gene IDs (updates 'weights' node attribute)}
\usage{
setUniverse(allIDs, data, minGenes, nCores = 1)
}
\arguments{
\item{allIDs}{character vector of gene IDs}

\item{data}{object containing GO graph}

\item{minGenes}{minimum number of genes attributed to GO term for this term
to be kept for the further statistical testing}

\item{nCores}{maximum number of processes}
}
\value{
updated data object containing GO graph
}
\description{
Helper function that removes genes from graph nodes based on the given set of
possible gene IDs (updates 'weights' node attribute)
}
