% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go-resources.R
\name{listGOAnnotations}
\alias{listGOAnnotations}
\title{List previously saved annotation files based on specified filter arguments}
\usage{
listGOAnnotations(
  species = NULL,
  goRelease = NULL,
  database = "ensembl",
  ensemblRelease = NULL,
  customName = NULL,
  path = NULL,
  returnLatest = FALSE
)
}
\arguments{
\item{species}{species identifier (example: 'hsapiens' for human)}

\item{goRelease}{GO release date in yyyy-mm-dd format (e.g., "2022-01-31")}

\item{database}{database used for acquiring of genesets. Only 'ensembl' is
currently supported}

\item{ensemblRelease}{numeric gene annotation version (e.g., 100 for Ensembl)}

\item{customName}{a string that will be matched as a substring against the names of
available custom annotation files}

\item{path}{path to the directory where the annotations files were saved}

\item{returnLatest}{logical value indicating whether to return information for the 
file with latest GO and Ensembl releases out of the ones matching the specified filter arguments.
If both standard and custom annotation files are among those matching specified filter arguments,
the information for the standard annotation with the latest GO and Ensembl releases is returned}
}
\value{
a data.frame containing information about previously saved annotations which
matched specified filter arguments
}
\description{
List previously saved annotation files based on specified filter arguments
}
