% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go-resources.R
\name{loadGOAnnotation}
\alias{loadGOAnnotation}
\title{Load file with latest release from previously retrieved GO annotation matching the specified
filter arguments}
\usage{
loadGOAnnotation(
  species = NULL,
  database = "ensembl",
  goRelease = NULL,
  ensemblRelease = NULL,
  customName = NULL,
  path = NULL
)
}
\arguments{
\item{species}{species identifier (example: 'hsapiens' for human)}

\item{database}{database used for acquiring of genesets. Only 'ensembl' is
currently supported}

\item{goRelease}{GO release date in yyyy-mm-dd format (e.g., "2022-01-31").}

\item{ensemblRelease}{gene annotation version (e.g., 100 for Ensembl).}

\item{customName}{a string that will be matched as a substring against the names of
available custom annotation files}

\item{path}{directory where evoGO annotations are stored. If NULL, the default package
directory is used ('extdata')}
}
\value{
list of evoGO objects containing the latest annotation (one per GO domain).
}
\description{
Load file with latest release from previously retrieved GO annotation matching the specified
filter arguments
}
