% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go-resources.R
\name{loadGOAnnotation}
\alias{loadGOAnnotation}
\title{Load previously saved GO annotation}
\usage{
loadGOAnnotation(
  species = NULL,
  database = "ensembl",
  goRelease = NULL,
  ensemblRelease = NULL,
  customName = NULL,
  path = NULL
)
}
\arguments{
\item{species}{species identifier (example: 'hsapiens' for human).}

\item{database}{database used for acquiring of genesets. Only 'ensembl' is
currently supported.}

\item{goRelease}{GO release date in yyyy-mm-dd format (e.g., "2022-01-31").}

\item{ensemblRelease}{gene annotation version (e.g., 100 for Ensembl).}

\item{customName}{a string that was used to store custom annotation file.}

\item{path}{directory where GO annotations are stored. If NULL, the default package
directory is used ('extdata').}
}
\value{
list of evoGO objects containing the latest annotation (one per GO domain).
}
\description{
Either a specific GO annotation can be loaded by using filter arguments or an
annotation will be selected based on the most recent available GO and Ensembl releases.
}
